mobile_app/Carbon-Footprint/requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.2.1
Pillow==10.0.0
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
